chore: rename file with LocalLayerItemRenderer
